Added a provider to manage yum repos the right way :)
I think we totally got it wrong the first time! Just deleting the file,
does not actually delete the repo correclty as it leaves the yum cache
unchanged.

So when I thought about how to solve that, I suddenly thought we should
just use the provider to rollback/delete/remove instead of just
managing the files. So this works very nicely and does exactly what one
would expect.

The strange/dangerous thing here is that I didn’t add a depend on the
yum cookbook. But I did that on purpose… I assume that if you want to
zap your obsolete repos, you need to define the once you want first so
the yum cookbook already is loaded for that.

I’ll see if I can add a check so it just gives some debug output is
it’s not actually there.