Added an option to call zap inline at a certain point in your run
This is needed for example when you clean out yum repos… As long as the
obsolete yum repos are still there ad somewhere along the line (in the
same run) you need a package from a new repo, it will fail because the
old (unavailable one) is still there. Now with the option immediately
you can call the zap action right where you need it, if it is important
for the remainder of your run.